requirements.txt
# Flask and extensions
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-Login==0.6.2
Flask-WTF==1.1.1
Flask-Mail==0.9.1
Flask-Migrate==4.0.4
Werkzeug==2.2.3
Jinja2==3.1.2
itsdangerous==2.1.2

# Database
SQLAlchemy==2.0.4
alembic==1.10.2

# Form handling and validation
WTForms==3.0.1
email-validator==2.0.0

# Environment variables
python-dotenv==1.0.0

# Date handling
pytz==2023.3

# Testing
pytest==7.3.1
pytest-flask==1.2.0

# Development tools
black==23.3.0
flake8==6.0.0

# Production server
gunicorn==20.1.0